+++ POI/trunk/org.openl.lib.poi35.dev.modified/src/org/apache/poi/hssf/record/formula/functions/MultiOperandNumericFunction.java
@@ -98,7 +98,7 @@ public abstract class MultiOperandNumericFunction implements FunctionWithArraySu
 
 	private static final int DEFAULT_MAX_NUM_OPERANDS = 30;
 
-	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 
 		double d;
 		try {
