+++ POI/trunk/org.openl.lib.poi35.dev.modified/src/org/apache/poi/hssf/record/formula/functions/NumericFunction.java
@@ -35,7 +35,7 @@ public abstract class NumericFunction implements FunctionWithArraySupport {
 	static final double TEN = 10.0;
 	static final double LOG_10_TO_BASE_e = Math.log(TEN);
 
-	protected static final double singleOperandEvaluate(ValueEval arg, int srcCellRow, short srcCellCol) throws EvaluationException {
+	protected static final double singleOperandEvaluate(ValueEval arg, int srcCellRow, int srcCellCol) throws EvaluationException {
 		ValueEval ve = OperandResolver.getSingleValue(arg, srcCellRow, srcCellCol);
 		double result = OperandResolver.coerceValueToDouble(ve);
 		checkValue(result);
@@ -48,7 +48,7 @@ public abstract class NumericFunction implements FunctionWithArraySupport {
 		}
 	}
 
-	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public final ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 		double result;
 		try {
 			result = eval(args, srcCellRow, srcCellCol);
@@ -68,7 +68,7 @@ public abstract class NumericFunction implements FunctionWithArraySupport {
 	}
 	// end changes ZS 
 
-	protected abstract double eval(ValueEval[] args, int srcCellRow, short srcCellCol) throws EvaluationException;
+	protected abstract double eval(ValueEval[] args, int srcCellRow, int srcCellCol) throws EvaluationException;
 
 	/* -------------------------------------------------------------------------- */
 	// intermediate sub-classes (one-arg, two-arg and multi-arg)
@@ -77,7 +77,7 @@ public abstract class NumericFunction implements FunctionWithArraySupport {
 		protected OneArg() {
 			// no fields to initialise
 		}
-		protected final double eval(ValueEval[] args, int srcCellRow, short srcCellCol) throws EvaluationException {
+		protected final double eval(ValueEval[] args, int srcCellRow, int srcCellCol) throws EvaluationException {
 			if (args.length != 1) {
 				throw new EvaluationException(ErrorEval.VALUE_INVALID);
 			}
@@ -91,7 +91,7 @@ public abstract class NumericFunction implements FunctionWithArraySupport {
 		protected TwoArg() {
 			// no fields to initialise
 		}
-		protected final double eval(ValueEval[] args, int srcCellRow, short srcCellCol) throws EvaluationException {
+		protected final double eval(ValueEval[] args, int srcCellRow, int srcCellCol) throws EvaluationException {
 			if (args.length != 2) {
 				throw new EvaluationException(ErrorEval.VALUE_INVALID);
 			}
@@ -109,7 +109,7 @@ public abstract class NumericFunction implements FunctionWithArraySupport {
 			_minArgs = minArgs;
 			_maxArgs = maxArgs;
 		}
-		protected final double eval(ValueEval[] args, int srcCellRow, short srcCellCol) throws EvaluationException {
+		protected final double eval(ValueEval[] args, int srcCellRow, int srcCellCol) throws EvaluationException {
 			int nArgs = args.length;
 			if (nArgs < _minArgs || nArgs > _maxArgs) {
 				throw new EvaluationException(ErrorEval.VALUE_INVALID);
